// fe/src/lib/api/generated/model/publicPlatformStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type PublicPlatformStatus = typeof PublicPlatformStatus[keyof typeof PublicPlatformStatus];


export const PublicPlatformStatus = {
  operational: 'operational',
  maintenance_scheduled: 'maintenance_scheduled',
  maintenance_active: 'maintenance_active',
} as const;